requirements.txt
streamlit
Pillow
pillow-avif-plugin
pillow-heif
